Allow fractional wait time between PDS jobs
This allows us to specify the wait time between PDS jobs in fractional
seconds to give us more flexibility in terms of controlling this.

+++ app/jobs/bulk_update_patients_from_pds_job.rb
@@ -11,7 +11,7 @@ class BulkUpdatePatientsFromPDSJob < ApplicationJob
     return unless Settings.pds.enqueue_bulk_updates
 
     patients = Patient.with_nhs_number.not_invalidated.not_deceased
-    wait_between_jobs = Settings.pds.wait_between_jobs.to_i
+    wait_between_jobs = Settings.pds.wait_between_jobs.to_f
 
     GoodJob::Bulk.enqueue do
       patients
